Delete, update product - admin function.

+++ confectionery/src/main/java/app/confectionery/product/repository/FileDataRepository.java
@@ -8,9 +8,11 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface FileDataRepository extends JpaRepository<FileData, Integer> {
+public interface FileDataRepository extends JpaRepository<FileData, Long> {
 
     Optional<FileData> findByName(String fileName);
     boolean existsByName(String name);
+
+
 
 }
